Move helpers to utils and add new cache factory
