+++ src/models/Usergroup.ts
@@ -2,12 +2,15 @@ import { model, SchemaDefinition, Model as M, SchemaTypes } from "mongoose";
 import { FLAG as UF, IUser } from "@models/User";
 import { ObjectId, existsValidator } from "@models/common";
 import { Base, IDoc, IDocRaw } from "./common";
+import newCache  = require("@utils/newCache");
 
 const Definition: SchemaDefinition = {
     name: { type: String, required: true }
 };
 
 export const FLAG = "usergroups";
+
+export const cache = newCache(FLAG);
 
 export interface IUsergroups extends IDocRaw {
     name: string;
