+++ src/modules/database/database.providers.ts
@@ -5,7 +5,7 @@ import { Model as UsersModel } from "@models/User";
 import { Model as UsergroupsModel } from "@models/Usergroup";
 import { Model as UserUsergroupsModel } from "@models/User-Usergroup";
 import { SystemService } from "@services/system";
-import { systemLogger } from "../common/helper/log";
+import { systemLogger } from "@utils/log";
 
 const getDatabaseUrl = () => {
     let host: string[] = [ ];
