+++ src/services/usergroups.service.ts
@@ -1,6 +1,8 @@
 import { Component, BadRequestException } from "@nestjs/common";
 import { Model as UsersModel } from "@models/User";
-import { Model as UsergroupsModel, IUsergroups } from "@models/Usergroup";
+import {
+    Model as UsergroupsModel, IUsergroups, cache
+} from "@models/Usergroup";
 import { Model as UserUsergroupsModel } from "@models/User-Usergroup";
 import { ObjectId } from "@models/common";
 import { BaseService } from "@services/base";
@@ -10,6 +12,7 @@ export class UsergroupsService extends BaseService<IUsergroups> {
 
     constructor() {
         super();
+        this.setCache(cache);
         this.setModel(UsergroupsModel);
     }
